requirements.txt
# Core AI/ML dependencies
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.4.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.23.0

# Face processing
insightface
onnxruntime-gpu

# Web framework
fastapi>=0.104.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.24.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0

# Optional optimization (comment out if not available)
# xformers>=0.0.22

# Clothing segmentation dependencies for undressing (uncomment when implementing)
# rembg>=2.0.50                    # Background and clothing removal for undressing
# u2net>=0.1.0                     # Universal segmentation for body exposure
# segment-anything>=1.0            # Meta's SAM for clothing segmentation

# Additional image processing (uncomment when implementing)
# scikit-image>=0.21.0             # Image processing algorithms for blending
# scipy>=1.11.0                    # Scientific computing for feathering and fusion

# Face restoration models (uncomment when implementing)
# gfpgan>=1.3.8                    # Face restoration (GFPGAN)
# basicsr>=1.4.2                   # Shared dependency for GFPGAN and RealESRGAN
# codeformer                       # Alternative face restoration (CodeFormer)

# Image upscaling models (uncomment when implementing)
# realesrgan>=0.3.0                # Real-ESRGAN for image super-resolution upscaling
